requirements.txt
langchain
langchain-community
langchain-groq
duckduckgo-search
ddgs
python-dotenv
wikipedia
arxiv
gradio
